Include descriptive commit statistics on non-tagged analysis
The data are (at least in principle) readily available, so
there's not reason to not include them in the commit list.

cmtList, surprisingly, is not a list, but a dictionary. The
nomenclature will be updated in a later commit. Also notice
that right now, we're duplicating work when the regular commit
list is computed after the file commit list. The information from
the first pass will need to be reused.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>