Convert the dispatcher to python
As usual, the bash shell script creates a magnitude more problems
for mildly complicated operations than is the case in a proper language.
Thus do the dispatching in python from now on.

This has the additional advantage that nothing else than a configuration
file is required to dispatch all analysis steps, which is mandatory
for the final product.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>