Make scatter plots robust against corner cases
When there are very few commits in a given revision range,
it can happen that all scattered points end up on the same x value,
which in turn causes an error from geom_smooth(). Don't run the
linear model in this case, also also remove any jitter.

NOTE: I've tried to replace a faulty graph with an error message
indicator graph, but this turns out to be soooooooo inconvenient
because of tons of side effects (the graph has to be evaluated once,
which causes problems in grid.arrange if the output has been set
before, and it's also nearly impossible to replace a graph out of
a list owing to the super-stupid tryCatch() mechanism of R -- who
would have thought that return() statements can ever be ignored
in programming languages...), I have decided to omit this feature.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>