Bring the multiple analysis methods together to share more of the same functions
The proximity based method was changed to use more of the
generic functions made in the prior commits. The main
differences have been reduced to only calculating the
different networks but share all other functions. The
only remaining differences are entirely due to the
different data aquired and not implementation differences

Signed-off-by: Mitchell Joblin <joblin.m@gmail.com>