an optimized version of the proximity based collaboration analysis is added as an alternative to the existing method
The aquisition and processing of the data needed to perform
the proximity based collaboration analysis turned out to be
extremely computationally intensive. The main contributing
factor was the fact that a git blame call was made for every
commit made during a revision range. The bulk of the optimization
is realized by only calling git blame once per file for the
latest file revision for the specified revision range. The
disadvantage of the optimization is if two commits are made
to the same line of code within the same revision range, only
the latests committed lines will be captured by the optimized
method.

The optimization was made as an alternative to the prior
method. If time is not an issue or the number of revisions
to anlayze is small, the non optimized method will give
more accurate results and should be used.

Signed-off-by: Mitchell Joblin <joblin.m@gmail.com>