we must consider the time of the commit in the fast version of the proximity based analysis
In the original version of the proximity based analysis
we had a snap shot of the file for every commit that
was made. Each snap shot represented a time slice.
In the fast version we only take one snap shot for
a single revision and infer all connections based
on that snap shot. To take into account the time aspect
we must use the commit time to figure out the order of
commits. If we didn't consider the time then all connections
inferred from the snap shot would be reciprocal and thus
defeat the purpose of a directed graph.

Signed-off-by: Mitchell Joblin <joblin.m@gmail.com>