add ctags functionality to improve proximity based analysis
The proximity based method was making the assumption that lines
of code within a given range (eg. +-20 LOC) were related. To
improve on this weak assumption ctags is utilized to identify
lines of code that are bracketed by a function. We then assume
that lines of code that exist in the same function are related
and therefore different developers making contributions to that
code are seen as collaborating.

Signed-off-by: Mitchell Joblin <joblin.m@gmail.com>