modification of collaboration metric for proximity based method
The collaboration metric prior to the ctags method
was to use the proximity (how close lines of codes
where placed to one another) to determine a weight.
The new ctags method simply gives a in equivalent
collaboration weight for all lines of code that fall
under the same function. Therefore two committer/authors
who make a contribution to the same function are given
a collaboration weight connection of 1. At the moment the
metric does not reflect the size of the contributions
only the frequency. We should consider whether to include
information about the size of the committed code. For example
a larger number of added lines could resulting in a larger
collaboration weight between to developers contributing to
the function.

Signed-off-by: Mitchell Joblin <joblin.m@gmail.com>