check if appropriate ctags is installed
A specific type of ctags (exuberant ctags) and verion
(v5.9) is absolutely necessary for the proximity based
analysis to operate. The commit implements a check whether
ctags exists and whether or not the correct version
is installed. If the correct ctags is not installed
or the incorrect version is found then the program
exits.

Signed-off-by: Mitchell Joblin <joblin.m@gmail.com>