Fill table release_timeline early
We know the release boundaries from the very beginning, albeit
without exact dates. Fill in the available information as early
as possible so that we can use it in the commit analysis steps.
Modify the time series code to update the data, not create new entries
(except for the release candidates).

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>