Refactor gen.iter.intervals
The function does two conceptionally different tasks: Firstly,
it constructs a list of intervals without regard to their content,
and then selects the non-empty portions by comparing the intervals
with the distribution of some content (dates). Split these task
into separate functions.

This way, we're also able to use the functionality on irregularly
spaced intervals as they appear when we consider per-release intervals.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>