Add per-release analysis pass
Some interesting time series analysis methods are performed
on a per-release basis, so add a processing step that does
exactly this. For now, start with an analysis that computes
how similar adjacent release time series were, and store the
results into the DB for later visualisation (or comparison among
projects).

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>