Functions to detrend time series
Add a helper function to detrend a time series by fitting
a linear model, and then using the residuals as new series
(see the comments on some statistical notes why this is not
overly accurate -- the method seems to work well, nontheless).
Also add a helper function th detrend all release ranges
of a project independently.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>