Extend cluster_user_mapping with page rank information
Page ranks are currently not stored in the database. As an
alternative to storing the page rank inside the cluster decomposition,
we could also have created a per release-range table with
entries for all persons contributing to the release, but since
prank is currently only needed when dealing with clusters, we
take this path.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>