Add routines to visualise collaboration clusters overviews
For the tasks of classifying/clustering collaboration clusters,
it is helpful to obtain an intuition what types appear for
a given project. Introduce a new file that acts as interactive
worksheet to produce these overviews.

Since data are only taken from the data base, the analysis can
commence independent of project processing.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>