Make main page rank table independent of clusterId
Calculating page ranks is done for every release range
and for every employed page rank technique. However, the
values are written before the clustering is even known, so
clusterId is not required in table pagerank.

To obtain the correct values in cluster_user_pagerank_view,
make sure that only the values for persons in the cluster
are included.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>