Remove page rank from per_cluster_statistics_view
The average page rank does not have any clear interpretation,
so get rid of it (it might be interesting to use some other
derived measure from the page rank distribution). Besides,
the view construction is not compatible with the current
definition of table page rank that does not contain a reference
to the release range.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>