Make ignoring merge commits optional
When time series boundaries are computed, the low-level commits
are queried directly from git. It can, however, happen that the
boundary coincides with a merge commit. If we ignore these commits,
the boundaries are not detected properly. Therefore allow to
take merges into account in special cases, and use the possibility
in getCommitDate.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>