Use release cycles as basis for ML analysis
With the infrastructure in place, query the release cycles from
the data base, and split the analysis accordingly.
For now, don't commence with the traditional perdiodic analysis,
although keep the code so we can re-instate it later -- it
could be useful for mining user instead of devel lists since the
former are usually only weakly coupled to the development cycles.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>