Write mailing-list activity data into the database
Populate the database with said data. Instead of writing the
raw series, we populate the database with the daily aggregated
transform, but don't compute any other smoothed representations -- this
can easily be done by the frontend with little computational effort.

In consequence, remove the local plotting functionality, and don't
produce static graphs anymore.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>