Use a revision unique name for temp file created for use of ctags
When processing revisions currently it can be the case that a conflict
arises during the blame analysis. Multiple process read and write to the
same file name in the same directory. The change prefixes the revision
onto the file name to prevent a conflict. Ideally we should use a totally
unique filename because running concurrent processing jobs with projects
in the same revision could still cause a conflict.

Signed-off-by: Mitchell Joblin <joblin.m@gmail.com>
Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>