Simplify network fixup
When NA values appear in the interest or communication networks,
we could not identify some elementary piece of information
regarding the author. Currently, we replace these NAs with
pseudo-names, but that will only cause interpretational problems
in later stages. Be decisive and eliminate these cases entirey
instead of processing semi-correct information.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>