Overhaul two-mode author-interest graph construction
Instead of producing a graph as (side-effect) result of the
author-interest computations, split the routine into two
parts: One to construct the graph, and one to write a SQL-friendly
representation of it into the database.

Instead of relying on snatm's custom-written plot routine to
layout the graph, we construct an igraph object with a minimal
set of parameters that are sufficient to reconstruct all visual
parameters later on, and write the result into the database.
gen.termplot is split into two new functions: construct.twomode.graph
and store.twomode.graph.

Callers in the analysis dispatcher are updated accordingly.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>