Add helper functions for working with in-DB two-mode graphs
Since two-mode graphs are now stored in the database, we need
to retrieve and process the data somehow. Two helper functions
for this purposes are provided in the new file twomode.r:
gen.twomode.graph() reconstructs an igraph object from the
information available in the database, and gen.twomode.dot()
prepares a visually appealing representation of the graph
in dot format.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>