Move initiate-response visualisation into a separate file
Since all required data are now stored in the database, we
can decouple visualisation from processing. Move the visualisation
functions from the analysis code into a separate file.

We use the opportunity to switch from log-scale to sqrt-scale
axes -- we deal with count data, which can be zero. This is not
optimal for log scale.

Besides, drop the non-transformed versions of the plots -- the
responses vary too much in size for this type of presentation
to be useful.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>