Unify tables mail_thread and thread_info
The tables are in a 1:1 relationship, and the information in
both tables is always available, so it does not seem like
a good choice to separate them.

Since there can be multiple mailing lists per project, we provide a
new field "ml" that is used to resolve to which list the data belong to.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>