Changed ID server return data format from json to jsonp
This allows us to run the ID servive on a different domain
than the website -- with regular json, the same origin policy
of most web browsers will prevent us from fetching data in this
case.

While this change is not important for obtaining IDs, it
is relevant when the non-ID API of the service is used
to, for instance, query lists of projects.

Signed-off-by: Oxana Ryashentseva <oxana.ryashentseva@siemens.com>
Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>