Add logger module for logging to console and files
The logger module instantiates a root log object for prosoda, which by
default has a console handler (i.e. log messages are printed to
stdout, with colour coding for convenience). The log level can be
specified on the command line. In addition, all log messages can be
logged into a file if desired. Unit and integration tests for the
logging and tests are also added in this commit.

Signed-off-by: Johannes Ebke <johannes.ebke.ext@siemens.com>
Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>