Add commit structure visualisation
The structure of commits is described by numerous covariates.
Perform multi-dimensional scaling and a principal component
analysis to find out how much of the actual parameter variance
can be explained by simpler models, and dynamically visualise
the results to see how the commit behaviour has changed oder
time. The visualisation allows us to obtain a notion of
"uniformity" among commits, which can be seen as another indicator
of engineering regularity.

When a project distinguished between RC and non-RC phases, we
separate the phases in the analysis.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>