R scripts now return nonzero error code on stop()
R scripts are now using a tryCatch construct with handlers for both
error and warning. The error handler dumps the stack frames and quits
with an error code, except if run in interactive mode. The warning
handler logs warnings to the debug logger.
The R process logging messages are now also shown on stdout/stderr.

Signed-off-by: Johannes Ebke <johannes.ebke.ext@siemens.com>
Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>