Improved error message for R errors
If an error in R code occurred, the error message was printed before the
traceback, which made it difficult to find. Now it is printed after the
traceback. In addition, there was a secondary bug in the python function
execute_command which further obscured the real error, which is now
fixed.

Signed-off-by: Johannes Ebke <johannes.ebke.ext@siemens.com>
Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>