Add labels to community visualization
The single most important person in each community is labeled
with the developer's name. All other developers are labeled with
their global id due to spacial constraints using names on each
node is not practical. For this reason we should provide a
legend mapping the id to a developer name.

Signed-off-by: Mitchell Joblin <mitchell.joblin.ext@siemens.com>