Introduce connection pool and fix concurrency
Up to now, one connection was used for communication with the MySQL
server. Apart from missing concurrency, this also led to the locking
being wholly ineffective, since any LOCK TABLE statement is valid for
all requests from this one connection.

This patch introduces the node-mysql connection pool, and gets a new
connection from the pool for each request. If a request necessitates
locking; the connection is held open and passed between callbacks so
that the lock remains in place. In addition, the user insertion and
update logic was simplified.

Signed-off-by: Johannes Ebke <johannes.ebke.ext@siemens.com>