Clear all clusters in db for given configuration before adding
Currently we clear each cluster id before adding a new cluster
however if the number of clusters is different between separate
runs of the analysis, then old clusters will not be cleared. Since
we use some non-deterministic clustering algorithms, it can be the
case that different numbers of clusters will be produced between
multiple runs of the algorithm.

Change:
- Clear all clusters for a given configuration (ie. project,
  release range, clustering method etc.) before adding new
  clusters

Signed-off-by: Mitchell Joblin <mitchell.joblin.ext@siemens.com>