Fix logic error in report generation batch job
Before, prosoda checked for dot files needed in report generation at
the time of batch job generation, before the analysis was run. This
change only adds the graphviz and lualatex runs after the analysis has
run. The BatchJob class was adapted to enable work items to be submitted
from subprocesses.

Signed-off-by: Johannes Ebke <johannes.ebke.ext@siemens.com>