Add widgets to display total diff size and relative diff size
Adds two widgets to the complexity pane, one that shows the sum of lines
in all diffs between two revisions, and one that shows the same number
divided by the lines of code in that revision, giving some indication
how much "churn" there is between revisions. OTOH, summarizing the
direct diff between two revisions might be a better metric for
stability, since changing the same lines over and over is not really
churn.

Signed-off-by: Johannes Ebke <johannes.ebke.ext@siemens.com>