Bugfix for Twitter Bootstrap style problem
When used inside of a Bootstrap Breadcrumb, dropdowns are not properly
displayed as a list. The reason is that the css style for class
"breadcrumb" says "display:inline-block" and this will allow several list
elements in dropdown displayed side by side, which is not desireable.
Setting the style of each dopdown-menu li element as display:inline will
remedy this.

Signed-off-by: Albert Eckert <albert.eckert@siemens.com>