Improve config file parsing error messages
When the command line options have been properly set,
but the configuration file cannot be parsed, it might be
perhaps helpful to tell an at least approximate hint to
the user. Re-stating the command line requirements (that
were correctly specified) without giving any further clue
is likely not the brightest option.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>