Replace nodejs* with idService*
It is not important in which language the ID service is implemented,
but more so that it delivers IDs. Consequently, replace all occurrences
of nodejsHostname and nodejsPort with idServiceHostname and
idServicePort to reflect this.

Patch created by

for file in `find . -name '*.py' -or -iname '*.r' | \
     xargs grep -l nodejs`; do \
sed -i -e 's/nodejs/idService/' $file; \
done

and some manual editing.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>