Use upstream version of package snatm
Some changes to package snatm are required for our purposes.
The changes have all been pushed upstream by now, so don't load
the package manually any more, but use the regular library mechanism.
To ensure that the fixes we need are available, explicitly check
that the installed version is at least 1.2.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>