Add a sloccount interface
sloccount can be used on project source trees to infer various
measures based on the COCOMO II model, in particular estimates
on the required development effort and its cost.
Additionally, it provides a guess of which languages the project
is composed of.

Introduce an interface to parse the output of sloccount and
present the data in a more regular form.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>