Add helper functions to deal with incremental time series
Some time series are expensive to compute, so we need ways to
update them incrementally when new release ranges are added
to a project (by default, most time series are computed during
the global analysis pass for the whole history). To enable
such updates of the database, add functions to query a plot ID
if it already exists, or add a new one if it doesn't.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>