Pre-process raw understand statistics before db serialisation
A typical understand run on an average to large project (for instance,
u-boot or qemu) produces nearly a million data points that have
previously been stored in table understand_raw. This is way too much
for a substantial history; reduce the data set to all pieces information
required to produce a boxplot.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>