Make index handling for clusters more explicit
We rely on the fact that vertices of clusters represented
in igraph objects have consecutive indices starting from 1,
but this is currently only true because of a side effeft (column
names get lost when the adjacency matrix is transposed in
performAnalysis). Since consecutive labels are a) assumed by
the rest of the code and b) required for some graph operations,
ensure that the property is actively enforced.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>