Fix handling invalid understand results
If understand does not produce results for one reason
or another, parse.understand() returns NULL. While the
subset selection on a NULL object succeeds, the test
for the number of columns does not -- it leads to a crash
instead of skipping this result set. Properly check
for invalid results by using a check for NULL, thus.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>