Integrate feature analysis by providing a CLI parameter.
This commit integrates the feature analysis added by the last commits.
Now the tagging configurations "feature" which is a project wide feature analysis and
"feature_file" for a per-file feature analysis are supported.
Additionally we introduce a --tagging parameter.
This parameter defaults to "default" which uses the tagging value from the configuration.
Sometimes you want to use the same configuration file but with different analysis methods, this could also be useful with CI later on.
Extracted LinkType class to an extra file so it can be used throughout the codebase (instead of magic strings).

Signed-off-by: Matthias Dittrich <matthi.d@gmail.com>
Reviewed-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>