Trace features and functions in the collaboration analysis.
We now support tracing commits, features and functions through the strengh of an relation.
This allows to include several new informations in the output.
Currently we output a new file 'adjacencyMatrix_max_weight.txt' which outputs the feature or
function with the most influence instead of the collaboration strength ('adjacencyMatrix.txt').
This would work for files as well but we have to change the 'File_Level' to something like '$filename' for the output to be usefull.

Signed-off-by: Matthias Dittrich <matthi.d@gmail.com>