Feature analysis edge cases, speedup and possibly incorrect data.
In the feature analysis it was possible to miss some early changes where one author would work alone on a set of lines.
Also the process can be made faster by calculating the feature groups for all commits.
Finally we handled some edge cases which are now required by the changed code.

Signed-off-by: Matthias Dittrich <matthi.d@gmail.com>