Improve the graph comparison code.
This commit improves the graph comparison code by not hoping for the analysis method to generate identical developer lists.
Instead the developer lists of different projects are merged by name and an mapping to local ids is generated out of this list.

This should ensure that for each graph the "correct" mapping is used to the new local ids.

Signed-off-by: Matthias Dittrich <matthi.d@gmail.com>