Change the generate_analysis_windows to be more stable.
The old implementation generates different analysis windows between calls.
The new implementation generates the same windows on every calls, it uses the date of the last commit as start reference.
An even more stable approach would be to use the date of the very first commit and go forward, this has the additional advantage that as new commits come in we only need to analyse the latest windows.

Signed-off-by: Matthias Dittrich <matthi.d@gmail.com>