Make dynamic graph tests externally available
"codeface dynamic" currently starts a server on localhost. Expose
the web service on port 0.0.0.0 so that it can be observed
on interfaces, and also forward the port from Vagrant to the
controlling host.

Notice that this does not mean that "codeface dynamic" is supposed
to be a production feature. This commit simplifies debugging.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@oth-regensburg.de>