Support LXC provisioning with Vagrant
As an alternative to VirtualBox, also support Linux
Containers (lxc) as Vagrant provisioner. This incurs
less overhead on Linux-based hosts. The provisioning
procedure is mostly identical, except for using a
different base image.

The means of allowing for multiple providers in a single
Vagrant file are admittedly a bit ugly, but support
for the problem currently seems sub-optimal upstream.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@oth-regensburg.de