Make ML author recovery diagnostics self explanatory
During mailing list analysis, we expect messages
to carry sender information in the form
"Author Name <author.email@domain.tld>
to identify  the respective persons correctly. In
many archive, bogus author sender specifications can
be found, and we try to correct them as good possible.
However, this is not clearly expressed in the
diagnostic message that's printed in this case.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>