Add web frontend widget for initate-response diagrams
Scatterplots to illuminate the mailing list communication
behaviour in terms of thread initiation and response
reception/generation have long been available in the codebase.
Display them in the web frontend.

Since this widget will serve as basis for a more detailled
documentation on how to create new widgets, the code will
be introduced in three parts. This one introduces a simple
version that does not distinguish between overview and
detail page.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@oth-regensburg.de>