Implement summary functions for init-response observations
Initiate-response diagrams are informative for a
given release cycle. Means of summarising the information
over the lifetime of a project are missing.

Try to fill this gap by providing a single-scalar summary
value for each release range. The methods contains several
parameters that will need to be adjusted to sensible
values.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@oth-regensburg.de>