Fix automatic download of mailinglist mbox files
The nntp-pull command needs to have the group and the mbox file-name
within single quotes, otherwise the download is written to the wrong
place.
As the command works again as expected, the system call in the R code
can be enabled again.

As a result, the download.r scripts downloads the mbox file to the given
download location automatically again.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>