Really re-install GitHub packages in packages.R
The current way of identification whether a package is installed or not
does not expect that a "wrong" version of the package is already
installed (e.g., from the package manager).

The new functions 'remove.installed.packages' and
'reinstall.package.from.github' do make sure that the previously
installed version of a package is indeed removed before installing it
from GitHub.

Before, there was the problem that 'tm.plugin.mail' has been installed
from the package manager (see the script
'integration-scripts/install_codeface_R.sh') and has NOT been
re-installed from GitHub, although the version from there is the correct
one.

[WM: Edit comments to use ## instead of #]
Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>
Reviewed-by: Wolfgang Mauerer <wolfgang.mauerer@oth-regensburg.de>