Added TODOs to require refactoring in project.py
Rationale for adding TODOs during cleanup:
The functions analyse_project and mailinglist_analyse both have
various issues which are non-trivial. This should and cannot be
fixed in this cleanup. Without adding a reminder we would risk
that this is never done and the codebase degrades.

This adding is supported both by the Google code standard and
good codestyle (e.g. Clean Code)