Fix bug regarding the most important node in a community
During clustering and appropriate visualization of developer networks,
it is necessary to find the most important node in a cluster (the one
with the highest rank, given a list of ranks). But, the current
implementation for identification of the most important node per
cluster actually only works for non-duplicate ranks.

If there are two nodes with the same rank which are in different
clusters and the most important node for their respective cluster, it
may be the case that the wrong one is picked. The new statements ensure
that exactly the one most important node from the current community is
selected, and not any node that has the same rank and, accidentally,
comes first in the node order.

Plus, remove trailing whitespace.

[CH: Adjust commit message.]
Signed-off-by: Thomas Bock <bockthom@fim.uni-passau.de>
Reviewed-by: Claus Hunsen <hunsen@fim.uni-passau.de>
Reviewed-by: Wolfgang Mauerer <wolfgang.mauerer@oth-regensburg.de>