Skip time-series analysis if too few releases are given
The time-series analysis prints an error message to the console if two
or less releases are present in the current configuration. However, the
analysis is continued and cosequently produces errors and stack-traces
in the console output.

To prevent obfuscated console output and confusion on the user side, the
time-series analysis now only logs a warning message and skips the
analysis if too few releases are given. As the previously given error
was not actually an error, this is the obvious solution.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>