Ensure that the log directory is writable
With LXC based deployments, the shared folder /vagrant
is not necessarily owned by user "vagrant" (this is only guaranteed
if the UID on the host machine is 1000, which is the case for
the user that was first created in most distributions). To
ensure that log files can be written, make the log/ directory
writable for everyone.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@oth-regensburg.de>