Fix commit_dependency table
Table identifiers are necessarily required in the
data model to refer to data objecs. Make table commit_dependency
adhere to this standard. This fixes some analysis errors.

NOTE: The only effective change in this patch is to set id
to NOT NULL; the large diff stems from format changes in the
MySQL workbench tool.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@oth-regensburg.de>