Fix timezone calculation
Rewrite timezone calculation such that relevant entries
in table commit can be changed with UPDATE statements.
The previously used INSERT INTO ... ON DUPLICATE KEY UPDATE
trickery does not work with recent mysql versions because
no commitHash is included in the statements; if an new entry would
be added with the statement, the NOT NULL constraint on the
commit ID could not be satisfied.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@oth-regensburg.de>