Increase travis timeout for R package installation
We need to compile loads of R packages from source, and this
can require more than 10 minutes on the travis-ci machines. This
leads to a timeout a a failing test build. Increase the time out
value.

The alternative would be, of course, to turn compilation
output back on again, but the countless C warnings from all the
dependent code are just too depressing to observe.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@oth-regensburg.de>