Provide diagnostic information for failing R tests
Same reason as the commit before. Decouple the R tests
further from the python testing infrastructure. In particular,
don't encode the tested R directories into the python
dispatcher. This especially allows us to spend more R
code for diagnostic messages.

This commit also fixes a problem with an invalid coercion
of the testthat summary results into a data.frame which is
not possible any more with the current version of testthat,
causing all R test to fail independent of any actual failures.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@oth-regensburg.de>