Add 'Base_Feature' analogous to 'File_Level'
For the 'feature' analysis, any change not occurring within an #ifdef
annotation is now mapped to the 'Base_Feature'. Accordingly, all changed
files of a commit are listed in the 'commit_dependency' table of the
database.

This change is in analogy to the 'File_Level' classification of the
'proximity' analysis, where any change not occurring within a function
is mapped to 'File_Level'.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>