Fix problems with NULL authors in ML analysis
In the mailing-list analysis, some persons with name 'NULL' could occur
in the database due to obscure patterns of names and mail addresses.

This patch fixes the parsing of two patterns:

1) only a mail address in angle brackets is provided: as the parsing result,
the first part of the mail address is returned as the author name.
(e.g., "From: <hans.huber@hubercorp.com')

2) a name looks like a mail address: as the parsing result, use only the
first part of that as name to avoid parsing problems in further steps of
the analysis and the ID service.
(e.g., "From: 'hans.huber@hubercorp.com' <hans.huber@hubercorp.com>")

[CH: Rephrase commit message and small fixes.]
Signed-off-by: Thomas Bock <bockthom@fim.uni-passau.de>
Reviewed-by: Claus Hunsen <hunsen@fim.uni-passau.de>
Reviewed-by: Wolfgang Mauerer <wolfgang.mauerer@oth-regensburg.de>