Improve parsing of date headers in the ML analysis
The default date format for parsing of mbox files is set by the 'snatm'
package to "%a, %d %b %Y %H %M:%S". In some projects' mbox files, the
weekday ("%a") is missing and the resulting date will get "NA" for those
e-mails.

This patch introduces a fixing routine which re-parses the date header
using the slightly adapted pattern "%d %b %Y %H:%M:%S" (omitted
weekday).

It might be the case, that the date is still set to NA for some mails,
but this patch catches, at least, the most common other pattern that has
been observed.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>