Improve parsing of date headers in the ML analysis
The default date format for parsing of mbox files is set by the 'snatm'
package to "%a, %d %b %Y %H %M:%S". In some projects' mbox files, e.g.,
the weekday ("%a") is missing and the resulting date will get "NA" for
those e-mails. Additionally, time-zone data (%z) have not been
incorporated (see original pattern below), although the commit analysis
incorporated this part of the date.

This patch introduces a fixing routine which re-parses ALL date headers
using a list of pre-defined patterns that have been sighted in real-world
.mbox files:
- "%a, %d %b %Y %H %M:%S" (the original pattern [1]),
- "%d %b %Y %H:%M:%S" (omitted weekday), and
- "%a, %d %b %Y %H:%M" (omitted seconds, "Wed, 21 Aug 2013 15:02 +0200").

[1] https://github.com/wolfgangmauerer/snatm/blob/master/pkg/R/makeforest.r#L47

The date headers are parsed right now as follows: first, trying to
include the time-zone information, then, excluding time-zone information
if not present. This guarantees that time-zone data is incorporated if
present, but potentially missing time-zone data does not break the date
parsing entirely for the current date header.

It might be the case, that the date is still set to NA for some mails,
but this patch catches, at least, the most common patterns that have
been observed.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>
Reviewed-by: Wolfgang Mauerer <wolfgang.mauerer@oth-regensburg.de>