Further improve parsing of date headers in the ML analysis
This patch extends the fixing routine which re-parses the date header to
fix the parsed date. To extend the existing code with another pattern,
the re-parsing is now augmented with the corresponding loop over all
defined patterns.

The newly introduced pattern is "%a, %d %b %Y %H:%M" (omitted seconds),
e.g., "Wed, 21 Aug 2013 15:02 +0200".

Fix trailing whitespace on the way.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>