Re-parse all date header in the ML analysis
This patch introduces functionality to re-parse ALL date headers while
parsing mails in the ML analysis. The re-parsing needs to be done,
because the date inside the given mbox file is initially parsed with the
pattern "%a, %d %b %Y %H:%M:%S" by the 'snatm' package - which does not
incorporate time-zone data (%z) [1], which is, on the other side,
incorporated in the commit analysis.

[1] https://github.com/wolfgangmauerer/snatm/blob/master/pkg/R/makeforest.r#L47

The date headers are parsed right now as follows: first, trying to
include the time-zone information, then, excluding time-zone information
if not present.
This guarantees that time-zone data is incorporated if present, but
potentially missing time-zone data does not break the date parsing
entirely for the current date header.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>