Further improve date parsing in ML analysis
Due to occurring pattern in real-world mbox files (e.g., QEMU from
GMane), the 'Date' header could not be parsed as it was preceeded by
"DATE:". The case-sensitive grep commands did not find those.
Consequently, the 'Date' header identification is changed to case-
insensitive.

Furthermore, if the 'Date' header is missing in a message, parsing for
this message is aborted early.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>
Reviewed-by: Wolfgang Mauerer <wolfgang.mauerer@oth-regensburg.de>