Add 'messageId' and 'creationDateOffset' columns to DB table 'mail'
For better traceability and to enable subsequent analyses, the message
ID of the mails is now stored in the database during the global ML
analysis.
To add the missing column in existing databases created with previous
schemas, run the following command inside MySQL:
> ALTER TABLE `codeface`.`mail` ADD `messageId` VARCHAR(255) NOT NULL
COMMENT '';

Furthermore, the time-zone offset of a mail's date is now stored in
the database during the global ML analysis.
To add the missing column in existing databases created with previous
schemas, run the following command inside MySQL:
> ALTER TABLE `codeface`.`mail` ADD `creationDateOffset` INT(11) NOT
NULL COMMENT '' AFTER `creationDate`;

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>
Reviewed-by: Wolfgang Mauerer <wolfgang.mauerer@oth-regensburg.de>
Acked-by: Wolfgang Mauerer <wolfgang.mauerer@oth-regensburg.de>