Fix cycle extraction in ML analysis
If there are no cycles (i.e., release ranges) in the database for the
given project - which happens in the ML test, for example -, the
subsetting of the result throws an error.

By, returning an empty data.frame instead, everything is fine. The user
only needs to check for "nrow(res) > 0" if they rely on the result.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>