Fix potential error with 'normalize' function in R
In the ML analysis, the function 'normalize' is called. Unfortunately,
the bare call is ambiguous as the function is defined in the 'snatm'
package (wanted) and the 'BiocGenerics' package (unwanted). Due to load
time, R might use the wrong function, throwing an error: "unable to find
an inherited method for function ‘normalize’ for signature ‘numeric’".

To avoid such an error, the function 'normalize' is now called with the
"snatm::" prefix.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>
Acked-by: Wolfgang Mauerer <wolfgang.mauerer@oth-regensburg.de>