Add option '--use-corpus' to ML analysis
To force Codeface to load a previously generated corpus at the start of
the ML analysis, the option '--use-corpus' is introduced to the Codeface
CLI and also the batch interface of the ML analysis.

The option '--use-corpus' is basically '!use.mbox' parameter to the
'codeface/R/ml/analysis.R::gen.forest' method.

If the option is not set, the corpus is (re-)generated from the .mbox
file by default.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>