Conway analysis: Add functions to download and parse jira issues
Using a portion of the Titan toolchain, add a pass that downloads
issues from a jira server, and create a data stucture in form of
a table that connects issues with commits.

This requires adding information about the jira URL, user credentials
and jira project names to the configuration files.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@oth-regensburg.de>
Signed-off-by: Carlos Andrade <carlos.andrade@acm.org>