Improve progress reporting
During release range resolved analysis, status messages are currently
based on the range under consideration. This is not particularly
helpful when a time-based approach is used; the range boundaries
are meaningless SHA hashes.

Explicitly print the index of the release range to provide a more
accurate impression of how much work is still left.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@oth-regensburg.de>