Remove support for reading Titan DSMs in Excel format
The Titan toolchain can produce DSMs in multiple formats. Since their
content is identical, it suffices to read the plain text variant.
Eliminate support for reading the Excel equivalent, since this
allows us to get rid of some package dependencies that are only
required for this purpose.

Use the opportunity to equip the file with license information.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@oth-regensburg.de>
Signed-off-by: Mitchell Joblin <mitchell.joblin.ext@siemens.com>