Make window size configuration accessible
Automatic analysis window generation operates on configurable
window sizes, but this capability is not exported to the configuration
file mechanism. Provide two new configuration parameters (windowSize
and numWindows) that allow for setting the temporal length of
analysis windows (in months), and the number of analysis windows
to use (-1 specifies to use all possible windows).

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@oth-regensburg.de>