Conway analysis: Exit quality analysis early if no motifs are found
When no motifs (or anti-motifs) are found in the graph, it does
not make much sense to perform a statistical quality analysis.
Exit early in this case.

TODO: This might actually be the wrong way to handle this -- it's
actually good, for instance, if there are no anti-motifs. Consider
this, and then rewrite the patch