Conway analysis: Use matching of induced subgraphs
Detecting socio-technical motifs in collaboration graphs is
currently performed by subgraph matching. However, we require the
motif to be an exact match, i.e., all edges present in the motif
must also be present in the larger graph. Change subgraph
matching to another method the can search for induced subgraphs.

TODO: This needs to be reviewed by someone else in addition to me.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@oth-regensburg.de>