Exit conway analysis early if there is no overlap between rcs and communication data
If not a single person contributed to the revision control system
and communicationd data, it does not make sense to try to
perform conway analysis. Stop the analysis early in this case
before running into errors.

Note that such a state is of course pathological; it is usually caused
when projects rewrite author names for one of both of the data sources
so that the ID service cannot disambiguate them properly any more.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@oth-regensburg.de>