Don't rely on ggplot2 internals to interleave lists
The R access system has been designed such that people stay out
of other package's internals beause they are not invited in, not
because the other package has a shotgun. But to proactively avoid
issues in the case ggplot2 should ever get hold of a shotgun
(or change its internal implementation), provide a proper public
implementation for interleaving lists instead of relying on the
unexported ggplot2:::interleave().

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@oth-regensburg.de>