Enhance db table issue_change to store comments
Table issue_comment stores issue ID, commenter ID and date of the comment addition. It does not store the comment text itself. Additionally, the data stored in the issue_comment table are virtually the same as in table issue_change and the table is therefore obsolete. Delete table issue_comment and add attribute for storing a comment to table issue_change. Thus it is possible to store either a comment added to an issue without any changes to the issue's field values, or one or more simultaneous changes of issue's field values with or without a comment (as multiple issue-tracking tools allow for).

Signed-off-by: Petr Pícha <ppicha@ntis.zcu.cz>