Create a relationship between issues and commits in DB
Many issue-tracking and VCS tools allow for an issue to mention
(or have another connection to) multiple commits, and for a commit
to link to multiple issues. And even if not specifically stated
in the tools the pairings between commits and issues (attained, e.g.,
by analysis) are useful for further analyses. Create a table in
the database to model this N:M relationship.

Signed-off-by: Petr Pícha <ppicha@ntis.zcu.cz>
Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@oth-regensburg.de>