Rename DB table cc_list to watcher_list
Database table cc_list stores people associated with an issue
(apart from the author and the asignee) who are to be informed
by email of any change applied to the issue. In common issue-tracking
tool terminology, these people are called "watchers" of the issue.
Rename table cc_list to watcher_list to better comply with the
common terminology.

Signed-off-by: Petr Pícha <ppicha@ntis.zcu.cz>
Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@oth-regensburg.de>