Remove Bugzilla specific, obsolete attributes from issue DB table
In the database table issue, attributes isRegression, subComponent and
subSubComponent are specific to Bugzilla and have no counterpart in
other issue-tracking tools. Furthermore, attribute modifiedDate is obsolete
since the date of the last modification can be obtained by querying the
latest item of the issue_change table related the this issue. Remove these
attributes. Value changes of the corresponding fields are still stored
through data in issue- and field_change tables.

Signed-off-by: Petr Pícha <ppicha@ntis.zcu.cz>
Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@oth-regensburg.de>