Rename attribute bugId in DB table issue to externalId
Attribute of table issue in the database storing the ID of
the issue in the source issue-tracking tool (server) is still
called bugId although the the table is not specifically designed
for bugs only but for issues generally. Renaming it to issueId
can result in misunderstanding that it stores Codeface ID of some
other issue. Rename the attribute to externalId to be less confusing
while referring to the ID of the issue in its source issue-tracking
tool (external to Codeface).

Signed-off-by: Petr Pícha <ppicha@ntis.zcu.cz>
Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@oth-regensburg.de>