Fix ML test by introducing a release range
As there are no cycles (i.e., release ranges) in the database for the
ML test project, the subsetting of the result in the function
'get.cycles.con' throws an error.

The problem is fixed by introducing a single revision range for the ML
test project and the needed functions for inserting everything into the
database.

Fix trailing whitespace along the way.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>