Downgrade MySQL to version 5.5 for TravisCI
There is weird stuff going on: TravisCI installs MySQL 5.6 by default,
but the Ubuntu package 'mysql-server' (and all analogous packages for
the client) wants to install version 5.5.

To prevent a clash, the installed version of TravisCI is removed and
version 5.5 is installed.

The installation script in the corresponding subfolder is untouched, so
that there is no problem with the default Vagrant containers.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>