Treat MySQL error 2013 in DB manager
The connection to the DB via the DB manager can fail not only with error
2006, but also with error 2013 "Lost connection to MySQL server during
query". To treat this kind of error now, an automatic reconnect is
attempted.

If the connection cannot be restored via a ping operation after error
2006 or 2013, the connection will automatically run into a timeout
later.

Reformat the whole file to fix indentation (remove tabs, fix indentation
size).

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>