Disable preprocessing in Doxygen
With 'ENABLE_PREPROCESSING = YES' in the doxygen.conf file, only that
annotated source code is incorporated in the Doxygen analysis for which
the presence condition is configured as true by default. Any other
variable code is not analyzed.
For example, the following function `foo` does not get indexed by
Doxygen if the constant `A` is not defined in the source code somewhere:
> #ifdef A
> function foo() { /*...*/ }
> #endif

To not omit such function definitions in the proximity analysis, the
preprocessing within Doxygen is disabled.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>