Add more correlation plots to global conway analysis
Try to find more insights into the data by providing a
scatter plots that compares #dev to normalised motif and anti-motif
counts, and include the motif ration in existing scatter plots.

Also remove the distinction between email and jira communication
plots -- the current plots can be used for either, and specialised
plots for the extra jira data can be added later on.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@oth-regensburg.de>