Re-engineer file-developer network construction in the Conway analysis
Generating the file-developer network is currently based on querying
the commit and commit_dependency tables. When an implementation
language is not supported by ctags (for instance, scala), the
analysis even at file level fails because the current method
constructs file contents from the changes to functions in a
file. Besides, the absolute size of files (and the LoC of
a file) can only be approximated with this method.

Rework the parsing mechanism by adding another round of
calls to git that explicitly determines the size of files,
and by then construction the information previously approximated
by dbmanager.py:get_file_dev() directly in conway.py:parseGitLogOutput().

Another implication of these changes is that not the full commit
message, but only the subject line of each commit are
searched for references to jira bug issues. This is the
originally intended mechanism since it is the convention
for Apache projects.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@oth-regensburg.de>